refactor api a bit
