+++ .travis.yml
@@ -1,7 +1,6 @@
 language: node_js
 node_js:
-  - '10'
-  - '8'
+  - '12'
 
 cache: yarn
